PLAT-2048: Allow custom AMI and user_data for managed nodegroups
A LaunchTemplate is created if both an ami_id and user_data are provided
for a managed nodegroup.

The user_data will go through minimal substituion so that psuedo
variables like AWS::Region and AWS::StackName can be filled in.

To allow the most basic bootstrap, the ClusterName is also provided.