Toggle nested stacks
You can only have 200 stacks per account, so having all modules
be nested stacks limits us to 28 deployments / aws account (or
at least unless you up the limits). It's mainly the eks module
that benefits from being separate (it hast a ton of resources),
and the toggle is probably more onerous to manage anyway, so we
don't support toggling that.