Modularize/Clean-up Provisioning (#25)
* Automatically setup autoscaler labels

* Checkpoint / cli skeleton

* Optimize-out per-az launchtemplate for managed

* Black formatting

* Questionable per-az lt optimize-out for unmanaged?

* Modularize s3 into nested stack

* Modularize all the things

* Move agent gen out of master stack+vpcazs

* Agent generator code moved out of the master stack
* VPC availability zone function moved into vpc stack,
  which is where it should have been to begin with

* Fix tests

* Toggle nested stacks

You can only have 200 stacks per account, so having all modules
be nested stacks limits us to 28 deployments / aws account (or
at least unless you up the limits). It's mainly the eks module
that benefits from being separate (it hast a ton of resources),
and the toggle is probably more onerous to manage anyway, so we
don't support toggling that.

* Cleanup

* Better nest toggling for s3

* Less janky nesting support, include eks toggle

* Reorganize provisioners

* Consolidate imports

* Remove cli stub test

* Modularize eks provisioning as well

* Remove WIP lambda code

This was going to replace the current thing, but hasn't been taken
up yet and is littering the code. Removing for now.