Normalize RC versions
Patch versions (the "-blah" part of "1.2.3-blah") have their hyphen
removed automatically during the python package build process, so
there's a mismatch between the terraform archive name (which uses
the version as-is) and the domino-cdk one (which has the truncated
version).

This makes the terraform archive match the domino_cdk one.

Note that this doesn't make a difference for release versions, just
patch versions.