Use api call k8s restriction
Only did it the other way because I was considering caching a json
of the data so boto/creds weren't a hard requirement, in which case
we'd need cache the version that has the full map for all eks k8s
versions. Since I'm not doing that, we can just get the data for
the version we're requesting.